+++ kadai3-2/yuonoda/downloader/downloader_test.go
@@ -2,7 +2,6 @@ package downloader_test
 
 import (
 	"os"
-	"reflect"
 	"testing"
 
 	downloader "github.com/yuonoda/gopherdojo-studyroom/kadai3-2/yuonoda/downloader"
@@ -55,41 +54,6 @@ func TestRun(t *testing.T) {
 			err = os.Remove(filePath)
 			if err != nil {
 				t.Error(err)
-			}
-		})
-	}
-
-}
-
-func TestFillByteArr(t *testing.T) {
-	cases := []struct {
-		name        string
-		arr         []byte
-		startAt     int
-		partArr     []byte
-		expectedArr []byte
-	}{
-		{
-			name:        "basic",
-			arr:         []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-			startAt:     3,
-			partArr:     []byte{4, 5, 6},
-			expectedArr: []byte{0, 0, 0, 4, 5, 6, 0, 0, 0, 0},
-		},
-		{
-			name:        "basic2",
-			arr:         []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-			startAt:     8,
-			partArr:     []byte{9, 10},
-			expectedArr: []byte{0, 0, 0, 0, 0, 0, 0, 0, 9, 10},
-		},
-	}
-
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			downloader.ExportedFillByteArr(c.arr[:], c.startAt, c.partArr)
-			if !reflect.DeepEqual(c.expectedArr, c.arr) {
-				t.Error("Array does not match")
 			}
 		})
 	}
