--- kadai3-2/yuonoda/downloader/export_test.go
@@ -1,3 +1,0 @@
-package downloader
-
-var ExportedFillByteArr = fillByteArr
